Fix 586 compiler issues.
Now that "-Werror" is turned on, 32-bit x86 compile problems are
fatal, so fix them.

*Should* be no functional change.

+++ usr/actor.c
@@ -239,7 +239,8 @@ actor_poll(void)
 		uint64_t time_left = actor_time_left(thread, current_time);
 		if (time_left) {
 			log_debug(7, "thread %08lx due %" PRIu64 ", wait %" PRIu64 " more",
-				  (long)thread, thread->ttschedule, time_left);
+				  (long)thread,
+				  (uint64_t)thread->ttschedule, time_left);
 
 			alarm(time_left);
 			break;
@@ -251,7 +252,7 @@ actor_poll(void)
 		log_debug(2, "thread %08lx was scheduled for "
 			  "%" PRIu64 ", curtime %" PRIu64 " q_forw %p "
 			  "&pend_list %p",
-			  (long)thread, thread->ttschedule,
+			  (long)thread, (uint64_t)thread->ttschedule,
 			  current_time, pend_list.next, &pend_list);
 
 		list_add_tail(&thread->list, &ready_list);
